pyproject.toml
[project]
name = "openview-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openreview-py>=1.54.1",
]
